refactor: ToggleSwitch -> Switch
